+++ Dyalect/Runtime/Types/Iterators/DyIteratorTypeInfo.cs
@@ -357,6 +357,12 @@ namespace Dyalect.Runtime.Types
             return (DyBool)res;
         }
 
+        private DyObject Contains(ExecutionContext ctx, DyObject self, DyObject item)
+        {
+            var seq = DyIterator.ToEnumerable(ctx, self);
+            return (DyBool)seq.Any(o => ctx.RuntimeContext.Types[o.TypeId].Eq(ctx, o, item).GetBool());
+        }
+
         protected override DyObject? InitializeInstanceMember(DyObject self, string name, ExecutionContext ctx) =>
             name switch
             {
@@ -380,6 +386,7 @@ namespace Dyalect.Runtime.Types
                 "reduce" => Func.Member(name, Reduce, -1, new Par("init", DyInteger.Zero), new Par("by")),
                 "any" => Func.Member(name, Any, -1, new Par("predicate")),
                 "all" => Func.Member(name, All, -1, new Par("predicate")),
+                "contains" => Func.Member(name, Contains, -1, new Par("value")),
                 _ => base.InitializeInstanceMember(self, name, ctx)
             };